+++ CHANGELOG.md
@@ -9,7 +9,11 @@ You can find and compare releases at the GitHub release page.
 ## [Unreleased]
 
 ### Added
+- Added exception if secret key or private/public key are missing
+
+### Fixed
 - Add command for generating certs
+
 
 ## [1.4.1] - 2022-01-24